pyproject.toml
[project]
name = "distillation-column-simulator"
version = "1.0.0"
description = "Advanced multi-stage continuous distillation column simulator with NRTL thermodynamic model"
authors = [
    {name = "Younes Ajeddig", email = "bouahligge@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["distillation", "chemical-engineering", "thermodynamics", "NRTL", "process-simulation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Fortran",
    "Topic :: Scientific/Engineering :: Chemistry",
]

dependencies = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pyyaml>=6.0.1",
    "pandas>=2.1.0",
    "numpy>=1.24.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.284",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/distillation-column-simulator"
Documentation = "https://github.com/YOUR_USERNAME/distillation-column-simulator#readme"
Repository = "https://github.com/YOUR_USERNAME/distillation-column-simulator"
Issues = "https://github.com/YOUR_USERNAME/distillation-column-simulator/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["python/distillation_wrapper"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.284",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
